[rajatbeck/AndroidTemplate#14] refactoring of packages
