+++ app/src/main/java/com/learn/androidtemplate/di/ViewModelFactoryModule.kt
@@ -1,12 +1,9 @@
 package com.learn.androidtemplate.di
 
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.learn.androidtemplate.home.HomeActivityViewModel
 import com.learn.androidtemplate.utils.ViewModelFactory
 import dagger.Binds
 import dagger.Module
-import dagger.multibindings.IntoMap
 
 @Module
 abstract class ViewModelFactoryModule{
