+++ app/src/main/java/com/learn/androidtemplate/di/home/HomeViewModelModule.kt
@@ -2,7 +2,7 @@ package com.learn.androidtemplate.di.home
 
 import androidx.lifecycle.ViewModel
 import com.learn.androidtemplate.di.ViewModelKey
-import com.learn.androidtemplate.home.HomeActivityViewModel
+import com.learn.androidtemplate.ui.home.HomeActivityViewModel
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
