+++ app/src/main/java/com/learn/androidtemplate/ui/Placeholder.kt
@@ -1,0 +1,2 @@
+package com.learn.androidtemplate.ui
+
