+++ app/src/main/java/com/learn/androidtemplate/ui/home/HomeAdapter.kt
@@ -1,0 +1,5 @@
+package com.learn.androidtemplate.ui.home
+
+class HomeAdapter{
+
+}
