+++ app/src/main/java/com/learn/androidtemplate/utils/Mapper.kt
@@ -1,12 +1,13 @@
-package com.learn.androidtemplate.repository
+package com.learn.androidtemplate.utils
 
 import com.learn.androidtemplate.data.FeedDTO
+import com.learn.androidtemplate.db.Feed
 
 fun FeedDTO.toDb() =
     Feed(
         id = this.id,
         user = this.user,
-        imageUrl = this.imageURL?:"",
+        imageUrl = this.imageURL ?: "",
         webFormatURL = this.webFormatURL,
         userId = this.userId,
         userImageURL = this.userImageURL,
